move player actions in src/features/player/actions
