+++ src/features/player/actions/walk-stairs.js
@@ -1,8 +1,8 @@
-import store              from '../../config/store';
-import generateMap        from '../../modules/generate-map';
-import showEndlessMessage from '../../modules/show-endless-message';
-import { uuidv4 }         from '../../modules/uuid-v4';
-import { SPRITE_SIZE }    from '../../config/constants';
+import store              from '../../../config/store';
+import generateMap        from '../../../modules/generate-map';
+import showEndlessMessage from '../../../modules/show-endless-message';
+import { uuidv4 }         from '../../../modules/uuid-v4';
+import { SPRITE_SIZE }    from '../../../config/constants';
 
 export default function walkStairs(nextTile, playerPos) {
   const { gameMode, randomMaps, floorNum, currentMap, storyMaps } = store.getState().world;
